@@ -1,11 +1,13 @@
-import ReactDOM from 'react-dom';
-import React from 'react';
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
 
 import App from './Components/App';
 
-ReactDOM.render(
-  <React.StrictMode>
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
+  <StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </StrictMode>
 );
